tests/gsim/check_attrel: renamed to check_gsim
